refactor basic 3d IM class
