+++ arm_robots/scripts/tutorial/ik_demo.py
@@ -1,156 +1,18 @@
 import argparse
 import pickle
-from typing import Callable, Optional
 
 from matplotlib import colors
 
 import rospy
 from arc_utilities import ros_init
 from arc_utilities.tf2wrapper import TF2Wrapper
+from arm_robots.basic_3d_pose_interactive_marker import Basic3DPoseInteractiveMarker
 from arm_robots.get_robot import get_moveit_robot
-from geometry_msgs.msg import Pose, Point
-from interactive_markers.interactive_marker_server import InteractiveMarkerServer
+from geometry_msgs.msg import Point
 from moveit_msgs.msg import RobotState, PlanningScene
 from sensor_msgs.msg import JointState
 from std_msgs.msg import ColorRGBA
-from visualization_msgs.msg import InteractiveMarker, InteractiveMarkerControl, Marker, InteractiveMarkerFeedback
-
-
-def make_box(scale: float):
-    marker = Marker(type=Marker.CUBE)
-    marker.scale.x = scale * 0.45
-    marker.scale.y = scale * 0.45
-    marker.scale.z = scale * 0.45
-    marker.color.r = 0.5
-    marker.color.g = 0.5
-    marker.color.b = 0.5
-    marker.color.a = 1.0
-
-    return marker
-
-
-def make_sphere(scale: float):
-    marker = Marker(type=Marker.SPHERE)
-    marker.scale.x = scale * 0.45
-    marker.scale.y = scale * 0.45
-    marker.scale.z = scale * 0.45
-    marker.color.r = 0.5
-    marker.color.g = 0.5
-    marker.color.b = 0.5
-    marker.color.a = 1.0
-
-    return marker
-
-
-def make_interactive_marker(name: str,
-                            position: Point,
-                            make_marker: Callable):
-    imarker = InteractiveMarker()
-    imarker.header.frame_id = "world"
-    imarker.pose.position = position
-    imarker.scale = 0.1
-
-    imarker.name = name
-
-    control = InteractiveMarkerControl()
-    control.always_visible = True
-    control.markers.append(make_marker(imarker.scale))
-    imarker.controls.append(control)
-
-    imarker.controls[0].interaction_mode = InteractiveMarkerControl.MOVE_ROTATE_3D
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 1
-    control.orientation.y = 0
-    control.orientation.z = 0
-    control.name = "rotate_x"
-    control.interaction_mode = InteractiveMarkerControl.ROTATE_AXIS
-    imarker.controls.append(control)
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 1
-    control.orientation.y = 0
-    control.orientation.z = 0
-    control.name = "move_x"
-    control.interaction_mode = InteractiveMarkerControl.MOVE_AXIS
-    imarker.controls.append(control)
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 0
-    control.orientation.y = 1
-    control.orientation.z = 0
-    control.name = "rotate_z"
-    control.interaction_mode = InteractiveMarkerControl.ROTATE_AXIS
-    imarker.controls.append(control)
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 0
-    control.orientation.y = 1
-    control.orientation.z = 0
-    control.name = "move_z"
-    control.interaction_mode = InteractiveMarkerControl.MOVE_AXIS
-    imarker.controls.append(control)
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 0
-    control.orientation.y = 0
-    control.orientation.z = 1
-    control.name = "rotate_y"
-    control.interaction_mode = InteractiveMarkerControl.ROTATE_AXIS
-    imarker.controls.append(control)
-
-    control = InteractiveMarkerControl()
-    control.orientation.w = 1
-    control.orientation.x = 0
-    control.orientation.y = 0
-    control.orientation.z = 1
-    control.name = "move_y"
-    control.interaction_mode = InteractiveMarkerControl.MOVE_AXIS
-    imarker.controls.append(control)
-    return imarker
-
-
-class Basic3DPoseInteractiveMarker:
-    def __init__(self,
-                 x: float = 0,
-                 y: float = 0,
-                 z: float = 0,
-                 shape: Optional[str] = None,
-                 make_marker: Optional[Callable] = None,
-                 name: str = 'basic_pose_marker'):
-        position = Point(x, y, z)
-        self.server = InteractiveMarkerServer(name)
-
-        if make_marker is None:
-            if shape is None:
-                raise NotImplementedError()
-            elif shape == 'sphere':
-                make_marker = make_sphere
-            elif shape == 'box':
-                make_marker = make_box
-            else:
-                raise NotImplementedError()
-
-        self.marker_name = 'imarker'
-        self.imarker = make_interactive_marker(self.marker_name, position, make_marker)
-        self.server.insert(self.imarker, self.on_feedback)
-
-        self.server.applyChanges()
-
-    def on_feedback(self, _: InteractiveMarkerFeedback):
-        self.server.applyChanges()
-
-    def set_pose(self, pose: Pose):
-        self.server.setPose(self.imarker.name, pose)
-        self.server.applyChanges()
-
-    def get_pose(self) -> Pose:
-        return self.server.get(self.marker_name).pose
+from visualization_msgs.msg import Marker
 
 
 @ros_init.with_ros("ik_demo")
@@ -225,7 +87,8 @@ def main():
 
         robot_state: RobotState
         if args.position_only:
-            robot_state = j.compute_collision_free_point_ik(default_robot_state, points, group_name, tip_names, scene_msg)
+            robot_state = j.compute_collision_free_point_ik(default_robot_state, points, group_name, tip_names,
+                                                            scene_msg)
         else:
             robot_state = j.compute_collision_free_pose_ik(default_robot_state, poses, group_name, tip_names, scene_msg)
